+++ puppyfw.php
@@ -7,7 +7,7 @@
  * Author URI:      https://truongwp.com/
  * Text Domain:     puppyfw
  * Domain Path:     /languages
- * Version:         0.4.2
+ * Version:         0.4.3
  *
  * @package         PuppyFW
  */